add run pipline module
